+++ lib/features/home/homepage.dart
@@ -4,7 +4,7 @@ import 'package:cherry_mvp/features/home/home_viewmodel.dart';
 import 'package:cherry_mvp/features/home/widgets/bottom_nav_bar.dart';
 import 'package:cherry_mvp/features/home/widgets/home_screen.dart';
 import 'package:cherry_mvp/features/messages/messagepage.dart';
-import 'package:cherry_mvp/features/profile/profilepage.dart';
+import 'package:cherry_mvp/features/profile/profile_page.dart';
 import 'package:cherry_mvp/features/search/search_page.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
